+++ src/main/java/gift/repository/MemberRepositoryInterface.java
@@ -16,7 +16,7 @@ public interface MemberRepositoryInterface {
 
     List<Product> findAllProductsFromWishListByEmail(String email);
 
-    void addProductToWishListByEmail(String email, Integer productId);
+    void addProductToWishListByEmail(String email, Long productId);
 
     boolean deleteProductFromWishListByEmail(String email, Long productId);
 }
